Added: Limit depth for the tree, Average calcualtion function
Can now controll the limit of the decision tree when changing the MAX_DEPTH variable.
It calculates the average between all nodes {0:n, 1:m} and makes a decsison based on all the nodes in the forest.
Original accuracy calculator takes the prediced value from each tree and wieghts if the forest had more True or False values